Move memory types to allocate/memory/types
